api/requirements.txt
asgiref==3.8.1
Django==5.1
sqlparse==0.5.1
djangorestframework==3.15.2
django-filter
django-cors-headers
django-environ==0.11.2
requests==2.32.3
# mysqlclient
